+++ bin/bench-prover/src/bench_functions.rs
@@ -38,7 +38,7 @@ pub fn setup_consume_note_with_new_account() -> Result<ExecutedTransaction> {
     let executed_transaction = mock_chain
         .build_tx_context(target_account.clone(), &[note.id()], &[])?
         .build()?
-        .execute()?;
+        .execute_blocking()?;
 
     // Apply delta to the target account to verify it is no longer new
     let target_account_after: Account = Account::from_parts(
@@ -83,7 +83,7 @@ pub fn setup_consume_multiple_notes() -> Result<ExecutedTransaction> {
         .build_tx_context(account.id(), &[note_1.id(), note_2.id()], &[])?
         .build()?;
 
-    let executed_transaction = tx_context.execute().unwrap();
+    let executed_transaction = tx_context.execute_blocking().unwrap();
 
     account.apply_delta(executed_transaction.account_delta()).unwrap();
     let resulting_asset = account.vault().assets().next().unwrap();