+++ crates/miden-testing/tests/scripts/send_note.rs
@@ -58,7 +58,7 @@ fn test_send_note_script_basic_wallet() -> anyhow::Result<()> {
         .tx_script(send_note_transaction_script)
         .extend_expected_output_notes(vec![OutputNote::Full(note)])
         .build()?
-        .execute()?;
+        .execute_blocking()?;
 
     // assert that the removed asset is in the delta
     let mut removed_assets: BTreeMap<_, _> = executed_transaction
@@ -122,6 +122,6 @@ fn test_send_note_script_basic_fungible_faucet() -> anyhow::Result<()> {
         .tx_script(send_note_transaction_script)
         .extend_expected_output_notes(vec![OutputNote::Full(note)])
         .build()?
-        .execute()?;
+        .execute_blocking()?;
     Ok(())
 }