+++ crates/miden-tx/src/host/mod.rs
@@ -1,4 +1,6 @@
 mod account_delta_tracker;
+use core::error::Error;
+
 use account_delta_tracker::AccountDeltaTracker;
 
 mod storage_delta_tracker;
@@ -19,7 +21,8 @@ mod tx_progress;
 use alloc::{boxed::Box, collections::BTreeMap, sync::Arc, vec::Vec};
 
 use miden_lib::transaction::{
-    TransactionEvent, TransactionEventError, TransactionKernelError,
+    TransactionEvent, TransactionEventData, TransactionEventError, TransactionEventHandling,
+    TransactionKernelError,
     memory::{CURRENT_INPUT_NOTE_PTR, NATIVE_NUM_ACCT_STORAGE_SLOTS_PTR},
 };
 use miden_objects::{
@@ -32,7 +35,7 @@ use miden_objects::{
 };
 pub use tx_progress::TransactionProgress;
 use vm_processor::{
-    ContextId, ErrorContext, ExecutionError, Felt, MastForest, MastForestStore, MemoryError,
+    AdviceMutation, ContextId, EventError, ExecutionError, Felt, MastForest, MastForestStore,
     ProcessState,
 };
 
@@ -153,121 +156,128 @@ where
     /// requested advice to the advice stack.
     pub fn handle_event(
         &mut self,
-        process: &mut ProcessState,
+        process: &ProcessState,
         transaction_event: TransactionEvent,
-        err_ctx: &impl ErrorContext,
-    ) -> Result<(), ExecutionError> {
+    ) -> Result<TransactionEventHandling, EventError> {
         // Privileged events can only be emitted from the root context.
         if process.ctx() != ContextId::root() && transaction_event.is_privileged() {
-            return Err(ExecutionError::event_error(
-                Box::new(TransactionEventError::NotRootContext(transaction_event as u32)),
-                err_ctx,
-            ));
+            return Err(Box::new(TransactionEventError::NotRootContext(transaction_event as u32)));
         }
 
-        match transaction_event {
-            TransactionEvent::AccountVaultBeforeAddAsset => Ok(()),
+        let advice_mutations = match transaction_event {
+            TransactionEvent::AccountVaultBeforeAddAsset => Ok(TransactionEventHandling::Handled(Vec::new())),
             TransactionEvent::AccountVaultAfterAddAsset => {
-                self.on_account_vault_after_add_asset(process)
+                self.on_account_vault_after_add_asset(process).map(|_| TransactionEventHandling::Handled(Vec::new()))
             },
 
-            TransactionEvent::AccountVaultBeforeRemoveAsset => Ok(()),
+            TransactionEvent::AccountVaultBeforeRemoveAsset => Ok(TransactionEventHandling::Handled(Vec::new())),
             TransactionEvent::AccountVaultAfterRemoveAsset => {
-                self.on_account_vault_after_remove_asset(process)
+                self.on_account_vault_after_remove_asset(process).map(|_| TransactionEventHandling::Handled(Vec::new()))
             },
 
-            TransactionEvent::AccountStorageBeforeSetItem => Ok(()),
+            TransactionEvent::AccountStorageBeforeSetItem => Ok(TransactionEventHandling::Handled(Vec::new())),
             TransactionEvent::AccountStorageAfterSetItem => {
-                self.on_account_storage_after_set_item(process)
+                self.on_account_storage_after_set_item(process).map(|_| TransactionEventHandling::Handled(Vec::new()))
             },
 
-            TransactionEvent::AccountStorageBeforeSetMapItem => Ok(()),
+            TransactionEvent::AccountStorageBeforeSetMapItem => Ok(TransactionEventHandling::Handled(Vec::new())),
             TransactionEvent::AccountStorageAfterSetMapItem => {
-                self.on_account_storage_after_set_map_item(process)
+                self.on_account_storage_after_set_map_item(process).map(|_| TransactionEventHandling::Handled(Vec::new()))
             },
 
             TransactionEvent::AccountBeforeIncrementNonce => {
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             },
             TransactionEvent::AccountAfterIncrementNonce => {
-                self.on_account_after_increment_nonce()
+                self.on_account_after_increment_nonce().map(|_| TransactionEventHandling::Handled(Vec::new()))
             },
 
             TransactionEvent::AccountPushProcedureIndex => {
-                self.on_account_push_procedure_index(process)
+                self.on_account_push_procedure_index(process).map(TransactionEventHandling::Handled)
             },
 
-            TransactionEvent::NoteBeforeCreated => Ok(()),
-            TransactionEvent::NoteAfterCreated => self.on_note_after_created(process),
+            TransactionEvent::NoteBeforeCreated => Ok(TransactionEventHandling::Handled(Vec::new())),
+            TransactionEvent::NoteAfterCreated => self.on_note_after_created(process).map(|_| TransactionEventHandling::Handled(Vec::new())),
 
-            TransactionEvent::NoteBeforeAddAsset => self.on_note_before_add_asset(process),
-            TransactionEvent::NoteAfterAddAsset => Ok(()),
+            TransactionEvent::NoteBeforeAddAsset => self.on_note_before_add_asset(process).map(|_| TransactionEventHandling::Handled(Vec::new())),
+            TransactionEvent::NoteAfterAddAsset => Ok(TransactionEventHandling::Handled(Vec::new())),
 
-            TransactionEvent::AuthRequest => self.on_signature_requested(process),
+            TransactionEvent::AuthRequest => {
+                let pub_key_hash = process.get_stack_word(0);
+                let message = process.get_stack_word(1);
+                let signature_key = Hasher::merge(&[pub_key_hash, message]);
+
+                let signature_opt = process
+                    .advice_provider()
+                    .get_mapped_values(&signature_key)
+                    .map(|slice| slice.to_vec());
+                let commitments_opt = process.advice_provider().get_mapped_values(&message).map(|slice| slice.to_vec());
+
+
+                Ok(TransactionEventHandling::Unhandled(TransactionEventData::AuthRequest { pub_key_hash, message, signature_key, signature_opt, commitments_opt }))
+            },
 
             TransactionEvent::PrologueStart => {
                 self.tx_progress.start_prologue(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             },
             TransactionEvent::PrologueEnd => {
                 self.tx_progress.end_prologue(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             },
 
             TransactionEvent::NotesProcessingStart => {
                 self.tx_progress.start_notes_processing(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             },
             TransactionEvent::NotesProcessingEnd => {
                 self.tx_progress.end_notes_processing(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             },
 
             TransactionEvent::NoteExecutionStart => {
-                let note_id = Self::get_current_note_id(process,err_ctx)?.expect(
+                let note_id = Self::get_current_note_id(process)?.expect(
                     "Note execution interval measurement is incorrect: check the placement of the start and the end of the interval",
                 );
                 self.tx_progress.start_note_execution(process.clk(), note_id);
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             },
             TransactionEvent::NoteExecutionEnd => {
                 self.tx_progress.end_note_execution(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             },
 
             TransactionEvent::TxScriptProcessingStart => {
                 self.tx_progress.start_tx_script_processing(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             }
             TransactionEvent::TxScriptProcessingEnd => {
                 self.tx_progress.end_tx_script_processing(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             }
 
             TransactionEvent::EpilogueStart => {
                 self.tx_progress.start_epilogue(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             }
             TransactionEvent::EpilogueEnd => {
                 self.tx_progress.end_epilogue(process.clk());
-                Ok(())
+                Ok(TransactionEventHandling::Handled(Vec::new()))
             }
             TransactionEvent::LinkMapSetEvent => {
-                LinkMap::handle_set_event(process)?;
-                Ok(())
+                return LinkMap::handle_set_event(process).map(TransactionEventHandling::Handled);
             },
             TransactionEvent::LinkMapGetEvent => {
-                LinkMap::handle_get_event(process)?;
-                Ok(())
+                return LinkMap::handle_get_event(process).map(TransactionEventHandling::Handled);
             },
             TransactionEvent::Unauthorized => {
               // Note: This always returns an error to abort the transaction.
               Err(self.on_unauthorized(process))
             }
         }
-        .map_err(|err| ExecutionError::event_error(Box::new(err),err_ctx))?;
+        .map_err(EventError::from)?;
 
-        Ok(())
+        Ok(advice_mutations)
     }
 
     /// Pushes a signature to the advice stack as a response to the `AuthRequest` event.
@@ -275,8 +285,8 @@ where
     /// The signature is fetched from the advice map and if it is not present, an error is returned.
     pub fn on_signature_requested(
         &mut self,
-        process: &mut ProcessState,
-    ) -> Result<(), TransactionKernelError> {
+        process: &ProcessState,
+    ) -> Result<Vec<AdviceMutation>, TransactionKernelError> {
         let pub_key = process.get_stack_word(0);
         let msg = process.get_stack_word(1);
 
@@ -285,12 +295,10 @@ where
         let signature = process
             .advice_provider()
             .get_mapped_values(&signature_key)
-            .map_err(|_| TransactionKernelError::MissingAuthenticator)?
+            .ok_or_else(|| TransactionKernelError::MissingAuthenticator)?
             .to_vec();
 
-        process.advice_provider_mut().stack.extend(signature);
-
-        Ok(())
+        Ok(vec![AdviceMutation::ExtendStack { values: signature }])
     }
 
     /// Creates a new [OutputNoteBuilder] from the data on the operand stack and stores it into the
@@ -351,11 +359,10 @@ where
     /// Expected stack state: [PROC_ROOT, ...]
     fn on_account_push_procedure_index(
         &mut self,
-        process: &mut ProcessState,
-    ) -> Result<(), TransactionKernelError> {
+        process: &ProcessState,
+    ) -> Result<Vec<AdviceMutation>, TransactionKernelError> {
         let proc_idx = self.acct_procedure_index_map.get_proc_index(process)?;
-        process.advice_provider_mut().push_stack(Felt::from(proc_idx));
-        Ok(())
+        Ok(vec![AdviceMutation::ExtendStack { values: vec![Felt::from(proc_idx)] }])
     }
 
     /// Handles the increment nonce event by incrementing the nonce delta by one.
@@ -528,7 +535,7 @@ where
     /// ```text
     /// [SALT, OUTPUT_NOTES_COMMITMENT, INPUT_NOTES_COMMITMENT, ACCOUNT_DELTA_COMMITMENT]
     /// ```
-    fn on_unauthorized(&self, process: &mut ProcessState) -> TransactionKernelError {
+    fn on_unauthorized(&self, process: &ProcessState) -> TransactionKernelError {
         let account_delta_commitment = process.get_stack_word(3);
         let input_notes_commitment = process.get_stack_word(2);
         let output_notes_commitment = process.get_stack_word(1);
@@ -551,10 +558,7 @@ where
     /// # Errors
     /// Returns an error if the address of the currently executing input note is invalid (e.g.,
     /// greater than `u32::MAX`).
-    fn get_current_note_id(
-        process: &ProcessState,
-        err_ctx: &impl ErrorContext,
-    ) -> Result<Option<NoteId>, ExecutionError> {
+    fn get_current_note_id(process: &ProcessState) -> Result<Option<NoteId>, EventError> {
         // get the note address in `Felt` or return `None` if the address hasn't been accessed
         // previously.
         let note_address_felt = match process.get_mem_value(process.ctx(), CURRENT_INPUT_NOTE_PTR) {
@@ -562,10 +566,9 @@ where
             None => return Ok(None),
         };
         // convert note address into u32
-        let note_address: u32 = note_address_felt.try_into().map_err(|_| {
-            ExecutionError::MemoryError(MemoryError::address_out_of_bounds(
-                note_address_felt.as_int(),
-                err_ctx,
+        let note_address = u32::try_from(note_address_felt).map_err(|_| {
+            Box::<dyn Error + Send + Sync>::from(format!(
+                "failed to convert {note_address_felt} into a memory address (u32)"
             ))
         })?;
         // if `note_address` == 0 note execution has ended and there is no valid note address