+++ crates/miden-tx/src/prover/prover_host.rs
@@ -1,13 +1,17 @@
 use alloc::{boxed::Box, sync::Arc, vec::Vec};
 
-use miden_lib::transaction::TransactionEvent;
+use miden_lib::transaction::{TransactionEvent, TransactionEventHandling};
 use miden_objects::{
     Word,
     account::{AccountDelta, PartialAccount},
+    assembly::{
+        SourceFile,
+        debuginfo::{Location, SourceSpan},
+    },
     transaction::{InputNote, InputNotes, OutputNote},
 };
 use vm_processor::{
-    BaseHost, ErrorContext, ExecutionError, MastForest, MastForestStore, ProcessState, SyncHost,
+    AdviceMutation, BaseHost, EventError, MastForest, MastForestStore, ProcessState, SyncHost,
 };
 
 use crate::{
@@ -68,7 +72,20 @@ where
 // HOST IMPLEMENTATION
 // ================================================================================================
 
-impl<STORE> BaseHost for TransactionProverHost<'_, STORE> where STORE: MastForestStore {}
+impl<STORE> BaseHost for TransactionProverHost<'_, STORE>
+where
+    STORE: MastForestStore,
+{
+    fn get_label_and_source_file(
+        &self,
+        _location: &Location,
+    ) -> (SourceSpan, Option<Arc<SourceFile>>) {
+        // For the prover, we assume that the transaction witness is a successfully executed
+        // transaction and so there should be no need to provide the actual source manager, as it
+        // is only used to improve error message quality which we shouldn't run into here.
+        (SourceSpan::UNKNOWN, None)
+    }
+}
 
 impl<STORE> SyncHost for TransactionProverHost<'_, STORE>
 where
@@ -80,13 +97,16 @@ where
 
     fn on_event(
         &mut self,
-        process: &mut ProcessState,
+        process: &ProcessState,
         event_id: u32,
-        err_ctx: &impl ErrorContext,
-    ) -> Result<(), ExecutionError> {
-        let transaction_event = TransactionEvent::try_from(event_id)
-            .map_err(|err| ExecutionError::event_error(Box::new(err), err_ctx))?;
+    ) -> Result<Vec<AdviceMutation>, EventError> {
+        let transaction_event = TransactionEvent::try_from(event_id).map_err(Box::new)?;
 
-        self.base_host.handle_event(process, transaction_event, err_ctx)
+        match self.base_host.handle_event(process, transaction_event)? {
+            TransactionEventHandling::Unhandled(_event_data) => {
+                self.base_host.on_signature_requested(process).map_err(EventError::from)
+            },
+            TransactionEventHandling::Handled(mutations) => Ok(mutations),
+        }
     }
 }